refactor: User, Login 패키지 User로 통합 및 API user 하위로 변경
