+++ src/main/java/upbrella/be/user/controller/LoginController.java
@@ -1,0 +1,13 @@
+package upbrella.be.user.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/oauth")
+@RequiredArgsConstructor
+public class LoginController {
+
+
+}
