+++ src/main/java/upbrella/be/user/controller/UserController.java
@@ -2,15 +2,17 @@ package upbrella.be.user.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import upbrella.be.rent.entity.History;
 import upbrella.be.rent.repository.RentRepository;
-import upbrella.be.umbrella.repository.UmbrellaRepository;
+import upbrella.be.user.dto.request.JoinRequest;
+import upbrella.be.user.dto.response.KakaoLoginResponse;
 import upbrella.be.user.dto.response.UmbrellaBorrowedByUserResponse;
 import upbrella.be.user.dto.response.UserInfoResponse;
+import upbrella.be.user.dto.token.KakaoOauthInfo;
+import upbrella.be.user.dto.token.OauthToken;
 import upbrella.be.user.entity.User;
+import upbrella.be.user.service.OauthLoginService;
 import upbrella.be.user.service.UserService;
 import upbrella.be.util.CustomResponse;
 
@@ -21,6 +23,9 @@ import javax.servlet.http.HttpSession;
 @RequiredArgsConstructor
 public class UserController {
 
+    private final OauthLoginService oauthLoginService;
+    private final UserService userService;
+    private final KakaoOauthInfo kakaoOauthInfo;
     private final RentRepository rentRepository;
 
     @GetMapping
@@ -78,5 +83,54 @@ public class UserController {
                         UmbrellaBorrowedByUserResponse.builder()
                                 .uuid(borrowedUmbrellaUuid)
                                 .build()));
+    }
+
+    @GetMapping("/login")
+    public ResponseEntity<CustomResponse> kakaoLogin(HttpSession session, String code) {
+
+        OauthToken kakaoAccessToken = oauthLoginService.getOauthToken(code, kakaoOauthInfo);
+        session.setAttribute("authToken", kakaoAccessToken);
+
+        KakaoLoginResponse kakaoLoggedInUser = oauthLoginService.processKakaoLogin(kakaoAccessToken.getAccessToken(), kakaoOauthInfo.getLoginUri());
+        long loginedUserId = userService.login(kakaoLoggedInUser.getId());
+        session.removeAttribute("authToken");
+
+        session.setAttribute("userId", loginedUserId);
+        return ResponseEntity
+                .ok()
+                .body(new CustomResponse<>(
+                        "success",
+                        200,
+                        "카카오 로그인 성공",
+                        null));
+    }
+
+    @PostMapping("/join")
+    public ResponseEntity<CustomResponse> kakaoJoin(HttpSession session, @RequestBody JoinRequest joinRequest) {
+
+        OauthToken kakaoAccessToken = (OauthToken) session.getAttribute("authToken");
+
+        if (session.getAttribute("userId") != null) {
+            throw new IllegalArgumentException("[ERROR] 이미 로그인된 상태입니다.");
+        }
+
+        if (kakaoAccessToken == null) {
+            throw new IllegalArgumentException("[ERROR] 로그인을 먼저 해주세요.");
+        }
+
+        KakaoLoginResponse kakaoLoggedInUser = oauthLoginService.processKakaoLogin(kakaoAccessToken.getAccessToken(), kakaoOauthInfo.getLoginUri());
+
+        long loginedUserId = userService.join(kakaoLoggedInUser.getId(), joinRequest);
+
+        session.removeAttribute("authToken");
+        session.setAttribute("userId", loginedUserId);
+
+        return ResponseEntity
+                .ok()
+                .body(new CustomResponse<>(
+                        "success",
+                        200,
+                        "카카오 회원가입 성공",
+                        null));
     }
 }
