+++ src/main/java/upbrella/be/user/exception/ExistingMemberException.java
@@ -1,4 +1,4 @@
-package upbrella.be.login.exception;
+package upbrella.be.user.exception;
 
 public class ExistingMemberException extends RuntimeException {
