+++ src/main/java/upbrella/be/user/service/UserService.java
@@ -2,8 +2,8 @@ package upbrella.be.user.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import upbrella.be.login.dto.request.JoinRequest;
-import upbrella.be.login.exception.ExistingMemberException;
+import upbrella.be.user.dto.request.JoinRequest;
+import upbrella.be.user.exception.ExistingMemberException;
 import upbrella.be.user.entity.User;
 import upbrella.be.user.repository.UserRepository;
